# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sabor & Gestión API
  description: API REST del sistema Sabor & Gestión para la gestión operativa del restaurante.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
...
paths:
...
  /api/platos:
    get:
      tags:
        - Platos
      summary: Listar catálogo de platos
      description: Endpoint público para consultar todos los platos del catálogo, con filtrado opcional por categoría.
      operationId: obtenerPlatos
      parameters:
        - name: category
          in: query
          required: false
          description: ID de la categoría para filtrar platos
          schema:
            type: string
          example: '{{categoriaId}}'
      responses:
        '200':
          description: Lista de platos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plato'
        '500':
          description: Error al obtener platos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    post:
      tags:
        - Platos
      summary: Crear un nuevo plato
      description: Crea un nuevo plato. Requiere rol Administrador. Obligatorio enviar imagenUrl e imagenPublicId obtenidas previamente desde POST /api/upload.
      operationId: crearPlato
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearPlatoRequest'
      responses:
        '201':
          description: Plato creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plato'
        '400':
          description: Falta imagenUrl o imagenPublicId de Cloudinary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al crear el plato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/platos/{id}:
    get:
      tags:
        - Platos
      summary: Obtener detalle de un plato
      description: Obtiene el detalle de un plato por ID con su categoría poblada. Requiere usuario autenticado.
      operationId: obtenerPlatoPorId
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del plato
          schema:
            type: string
          example: '{{platoId}}'
      responses:
        '200':
          description: Detalle del plato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plato'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Plato no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al obtener el plato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    put:
      tags:
        - Platos
      summary: Actualizar un plato
      description: Actualiza los campos de un plato. Requiere rol Administrador. Si se reemplaza la imagen, elimina automáticamente la anterior de Cloudinary.
      operationId: actualizarPlato
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del plato
          schema:
            type: string
          example: '{{platoId}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarPlatoRequest'
      responses:
        '200':
          description: Plato actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plato'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Plato no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al actualizar el plato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    delete:
      tags:
        - Platos
      summary: Eliminar un plato e imagen
      description: Elimina el plato de la base de datos y su archivo asociado en Cloudinary. Requiere rol de Administrador.
      operationId: eliminarPlato
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del plato
          schema:
            type: string
          example: '{{platoId}}'
      responses:
        '200':
          description: Plato e imagen eliminados correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeRespuestaResponse'
              example:
                mensaje: Plato e imagen eliminados correctamente
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Plato no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al eliminar el plato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
...
  /api/ubicaciones:
    get:
      tags:
        - Ubicaciones
      summary: Listar ubicaciones físicas
      description: Obtiene la lista de todas las ubicaciones o áreas registradas (ej. Salón Principal, Terraza, Patio). Requiere usuario autenticado.
      operationId: obtenerUbicaciones
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de ubicaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ubicacion'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al obtener ubicaciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    post:
      tags:
        - Ubicaciones
      summary: Crear una nueva ubicación
      description: Registra una nueva área física en el restaurante. Requiere rol Administrador. Valida unicidad de nombre.
      operationId: crearUbicacion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearUbicacionRequest'
      responses:
        '201':
          description: Ubicación creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ubicacion'
        '400':
          description: Nombre requerido o ya existente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al crear ubicación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/ubicaciones/{id}:
    put:
      tags:
        - Ubicaciones
      summary: Actualizar una ubicación
      description: Actualiza el nombre o descripción de una ubicación existente. Requiere rol Administrador.
      operationId: actualizarUbicacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la ubicación
          schema:
            type: string
          example: '{{ubicacionId}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearUbicacionRequest'
      responses:
        '200':
          description: Ubicación actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ubicacion'
        '400':
          description: Nombre requerido o ubicación ya existente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Ubicación no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al actualizar ubicación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    delete:
      tags:
        - Ubicaciones
      summary: Eliminar una ubicación
      description: Elimina una ubicación por ID. Requiere rol Administrador.
      operationId: eliminarUbicacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la ubicación
          schema:
            type: string
          example: '{{ubicacionId}}'
      responses:
        '200':
          description: Ubicación eliminada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeRespuestaResponse'
              example:
                mensaje: Ubicación eliminada correctamente
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Ubicación no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al eliminar ubicación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Autenticación mediante token JWT en cabecera: Authorization: Bearer <token>'
  schemas:
    MensajeErrorResponse:
      type: object
      required:
        - mensaje
      properties:
        mensaje:
          type: string
    MensajeRespuestaResponse:
      type: object
      required:
        - mensaje
      properties:
        mensaje:
          type: string
          example: Operación realizada exitosamente
...
    Categoria:
      type: object
      properties:
        id:
          type: string
          example: '000000000000000000000010'
        _id:
          type: string
          example: '000000000000000000000010'
        nombre:
          type: string
          example: Platos Fuertes
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Plato:
      type: object
      properties:
        id:
          type: string
          example: '000000000000000000000020'
        _id:
          type: string
          example: '000000000000000000000020'
        nombre:
          type: string
          example: Lomo Montado
        descripcion:
          type: string
          example: Corte de lomo vacuno acompañado con papas fritas, arroz y huevo frito.
        precio:
          type: number
          example: 45
        imagenUrl:
          type: string
          example: https://res.cloudinary.com/demo/image/upload/lomo.jpg
        imagenPublicId:
          type: string
          example: platos/lomo_montado_123
        disponible:
          type: boolean
          example: true
        categoria:
          oneOf:
            - type: string
              example: '000000000000000000000010'
            - $ref: '#/components/schemas/Categoria'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CrearPlatoRequest:
      type: object
      required:
        - nombre
        - descripcion
        - precio
        - imagenUrl
        - imagenPublicId
        - categoria
      properties:
        nombre:
          type: string
          example: Pique Macho
        descripcion:
          type: string
          example: Trozos de lomo salteado con salchichas, cebolla, locoto y papas fritas.
        precio:
          type: number
          example: 65
        imagenUrl:
          type: string
          example: https://res.cloudinary.com/demo/image/upload/pique.jpg
        imagenPublicId:
          type: string
          example: platos/pique_123
        categoria:
          type: string
          example: '{{categoriaId}}'
    ActualizarPlatoRequest:
      type: object
      properties:
        nombre:
          type: string
          example: Pique Macho Especial
        descripcion:
          type: string
          example: Pique tradicional con porción extra de carne.
        precio:
          type: number
          example: 75
        imagenUrl:
          type: string
          example: https://res.cloudinary.com/demo/image/upload/pique_esp.jpg
        imagenPublicId:
          type: string
          example: platos/pique_esp_456
        categoria:
          type: string
          example: '{{categoriaId}}'
        disponible:
          type: boolean
          example: true
...
    Ubicacion:
      type: object
      properties:
        id:
          type: string
          example: '000000000000000000000040'
        nombre:
          type: string
          example: Terraza
    CrearUbicacionRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          example: Patio Central
        descripcion:
          type: string
          example: Área al aire libre con mesas con sombrilla
...